+++ asdf/tests/test_fits_embed.py
@@ -50,7 +50,13 @@ def create_asdf_in_fits():
 
     return fits_embed.AsdfInFits(hdulist, tree)
 
-def test_embed_asdf_in_fits_file(tmpdir):
+# Testing backwards compatibility ensures that we can continue to read and
+# write files that use the old convention of ImageHDU to store the ASDF file.
+@pytest.mark.parametrize('backwards_compat', [False, True])
+def test_embed_asdf_in_fits_file(tmpdir, backwards_compat):
+    fits_testfile = str(tmpdir.join('test.fits'))
+    asdf_testfile = str(tmpdir.join('test.asdf'))
+
     hdulist = fits.HDUList()
     hdulist.append(fits.ImageHDU(np.arange(512, dtype=np.float), name='SCI'))
     hdulist.append(fits.ImageHDU(np.arange(512, dtype=np.float), name='DQ'))
@@ -69,24 +75,30 @@ def test_embed_asdf_in_fits_file(tmpdir):
     }
 
     ff = fits_embed.AsdfInFits(hdulist, tree)
-    ff.write_to(os.path.join(str(tmpdir), 'test.fits'))
+    ff.write_to(fits_testfile, use_image_hdu=backwards_compat)
 
-    ff2 = asdf.AsdfFile(tree)
-    ff2.write_to(os.path.join(str(tmpdir), 'plain.asdf'))
-
-    with fits.open(os.path.join(str(tmpdir), 'test.fits')) as hdulist2:
+    with fits.open(fits_testfile) as hdulist2:
         assert len(hdulist2) == 3
         assert [x.name for x in hdulist2] == ['SCI', 'DQ', 'ASDF']
         assert_array_equal(hdulist2[0].data, np.arange(512, dtype=np.float))
-        assert hdulist2['ASDF'].data.tostring().strip().endswith(b"...")
+        asdf_hdu = hdulist2['ASDF']
+        assert asdf_hdu.data.tostring().startswith(b'#ASDF')
+        # When in backwards compatibility mode, the ASDF file will be contained
+        # in an ImageHDU
+        if backwards_compat:
+            assert isinstance(asdf_hdu, fits.ImageHDU)
+            assert asdf_hdu.data.tostring().strip().endswith(b'...')
+        else:
+            assert isinstance(asdf_hdu, fits_embed._AsdfHDU)
+            assert len(asdf_hdu.data) % 2880 == 0
 
         with fits_embed.AsdfInFits.open(hdulist2) as ff2:
             assert_tree_match(tree, ff2.tree)
 
             ff = asdf.AsdfFile(copy.deepcopy(ff2.tree))
-            ff.write_to('test.asdf')
+            ff.write_to(asdf_testfile)
 
-    with asdf.AsdfFile.open('test.asdf') as ff:
+    with asdf.AsdfFile.open(asdf_testfile) as ff:
         assert_tree_match(tree, ff.tree)
 
 
@@ -101,7 +113,10 @@ def test_embed_asdf_in_fits_file_anonymous_extensions(tmpdir):
     with fits.open(os.path.join(str(tmpdir), 'test.fits')) as hdulist:
         assert len(hdulist) == 4
         assert [x.name for x in hdulist] == ['PRIMARY', '', '', 'ASDF']
-        assert hdulist['ASDF'].data.tostring().strip().endswith(b"...")
+        asdf_hdu = hdulist['ASDF']
+        assert isinstance(asdf_hdu, fits_embed._AsdfHDU)
+        assert asdf_hdu.data.tostring().startswith(b'#ASDF')
+        assert len(asdf_hdu.data) % 2880 == 0
 
         with fits_embed.AsdfInFits.open(hdulist) as ff2:
             assert_tree_match(asdf_in_fits.tree, ff2.tree)
@@ -111,6 +126,60 @@ def test_embed_asdf_in_fits_file_anonymous_extensions(tmpdir):
 
     with asdf.AsdfFile.open('test.asdf') as ff:
         assert_tree_match(asdf_in_fits.tree, ff.tree)
+
+
+@pytest.mark.xfail(
+    reason="In-place update for ASDF-in-FITS does not currently work")
+def test_update_in_place(tmpdir):
+    tempfile = str(tmpdir.join('test.fits'))
+
+    # Create a file and write it out
+    asdf_in_fits = create_asdf_in_fits()
+    asdf_in_fits.write_to(tempfile)
+
+    # Open the file and add data so it needs to be updated
+    with fits_embed.AsdfInFits.open(tempfile) as ff:
+        ff.tree['new_stuff'] = "A String"
+        ff.update()
+
+    # Open the updated file and make sure everything looks okay
+    with fits_embed.AsdfInFits.open(tempfile) as ff:
+        assert ff.tree['new_stuff'] == "A String"
+        assert_tree_match(ff.tree['model'], asdf_in_fits.tree['model'])
+
+
+def test_update_and_write_new(tmpdir):
+    tempfile = str(tmpdir.join('test.fits'))
+    newfile = str(tmpdir.join('new.fits'))
+
+    # Create a file and write it out
+    asdf_in_fits = create_asdf_in_fits()
+    asdf_in_fits.write_to(tempfile)
+
+    # Open the file and add data so it needs to be updated
+    with fits_embed.AsdfInFits.open(tempfile) as ff:
+        ff.tree['new_stuff'] = "A String"
+        ff.write_to(newfile)
+
+    # Open the updated file and make sure everything looks okay
+    with fits_embed.AsdfInFits.open(newfile) as ff:
+        assert ff.tree['new_stuff'] == "A String"
+        assert_tree_match(ff.tree['model'], asdf_in_fits.tree['model'])
+
+
+@pytest.mark.xfail(
+    reason="ASDF HDU implementation does not currently reseek after writing")
+def test_access_hdu_data_after_write(tmpdir):
+    # There is actually probably not a great reason to support this kind of
+    # functionality, but I am adding a test here to record the failure for
+    # posterity.
+    tempfile = str(tmpdir.join('test.fits'))
+
+    asdf_in_fits = create_asdf_in_fits()
+    asdf_in_fits.write_to(tempfile)
+    asdf_hdu = asdf_in_fits._hdulist['ASDF']
+
+    assert asdf_hdu.data.tostring().startswith('#ASDF')
 
 
 def test_create_in_tree_first(tmpdir):